coding-pr-agent/requirements.txt
python-dotenv>=1.0.0
e2b-code-interpreter>=1.5.0
openai>=1.0.0
rich>=13.0.0 